implement the view updateranking

+++ pes/models.py
@@ -70,8 +70,8 @@ class Match(models.Model):
     teamB = models.ForeignKey(Team, related_name='+')
     goalA = models.IntegerField(blank=True, null=True)
     goalB = models.IntegerField(blank=True, null=True)
-    deltaA = models.IntegerField(blank=True, null=True)
-    deltaB = models.IntegerField(blank=True, null=True)
+    deltaA = models.FloatField(blank=True, null=True)
+    deltaB = models.FloatField(blank=True, null=True)
 
     def play(self):
         a, b = self.goalA, self.goalB
